chore: move hbar limiter tests to separate job
